requirements_api.txt
fastapi==0.66.0
alembic~=1.7
graphene==3.0b7
cryptography==3.4.6
git+https://github.com/polkascan/graphene-sqlalchemy.git
PyMySQL==1.0.2
python-dotenv==0.15.0
pytest==6.2.2
sentry-sdk==0.19.5
SQLAlchemy==1.3.22
starlette==0.14.2
starlette-graphene3==0.5.0
tenacity==6.3.1
uvicorn[standard]==0.13.3
graphql-core==3.1.2
graphql-relay==3.1.0
websockets==8.1
git+https://github.com/polkascan/graphene-sqlalchemy-filter.git
substrate-interface~=1.1.7
broadcaster==0.2.0
asyncio-redis==0.16.0
sqlakeyset==1.0.1611286921
git+https://github.com/polkascan/sqlalchemy-pagination.git@f8ae4b458b345dc0a2ec09c496c6d204505dda50

opentelemetry-distro==0.29b0
opentelemetry-exporter-otlp==1.10.0
opentelemetry-instrumentation-fastapi==0.29b0
